Addng .zshrc and refactoring aliases and paths

+++ install.sh
@@ -6,5 +6,8 @@ for f in * .*; do
     if [ "$f" = ".git" ]; then
         continue
     fi
-    ln -s ~/.dotfiles/$f ~/$f
+
+    if [ ! -e ~/$f ]; then
+        ln -s ~/.dotfiles/$f ~/$f
+    fi
 done
